refine adaptor naming convention
