+++ Src/Extern/SENSEI/AMReX_AmrParticleDataAdaptor.H
@@ -4,19 +4,18 @@
 // AMReX includes
 #include <AMReX_Config.H>
 #ifdef AMREX_PARTICLES
+
 #include <AMReX_Particles.H>
 
 #include <AMReX_AmrMesh.H>
 #include <AMReX_MultiFab.H>
-
+#include <AMReX_AmrDataAdaptor.H>
+#include <AMReX_ParticleDataAdaptor.H>
 // sensei includes
 #include "DataAdaptor.h"
 
-
-class vtkPolyData;
 namespace amrex
 {
-
 template<int NStructReal, int NStructInt, int NArrayReal, int NArrayInt>
 class AmrParticleDataAdaptor : public sensei::DataAdaptor
 {
@@ -24,44 +23,14 @@ public:
   static AmrParticleDataAdaptor* New();
   senseiTypeMacro(AmrParticleDataAdaptor, sensei::DataAdaptor);
 
-  // set the source of the simulation data
   int SetDataSource(
+    Amr *amr,
     amrex::ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt> * particles,
     const std::map<std::string, std::vector<int>> & rStructs = {},
     const std::map<std::string, int> & iStructs = {},
     const std::map<std::string, std::vector<int>> & rArrays = {},
     const std::map<std::string, int> & iArrays = {}
   );
-
-  // set the data array names maps internally
-  int SetArrayNames(
-    const std::map<std::string, std::vector<int>> & rStructs,
-    const std::map<std::string, int> & iStructs,
-    const std::map<std::string, std::vector<int>> & rArrays,
-    const std::map<std::string, int> & iArrays);
-
-  // force mesh origin to 0,0,0. off by default
-  void SetPinMesh(int val);
-
-  // get particle id numbers
-  int AddParticlesIDArray(vtkDataObject* mesh);
-
-  // get particle cpu numbers (process each particle was generated on)
-  int AddParticlesCPUArray(vtkDataObject* mesh);
-
-  // get particle integer arrays in Structs of Arrays format
-  int AddParticlesSOAIntArray(const std::string &arrayName, vtkDataObject* mesh);
-
-  // get particle real arrays in Structs of Arrays format
-  int AddParticlesSOARealArray(const std::string &arrayName, vtkDataObject* mesh);
-
-  // get particle integer arrays in Array Of Structs format
-  int AddParticlesAOSIntArray(const std::string &arrayName, vtkDataObject* mesh);
-
-  // get particle real arrays in Array Of Structs format
-  int AddParticlesAOSRealArray(const std::string &arrayName, vtkDataObject* mesh);
-
-
 
   // SENSEI API
 #if SENSEI_VERSION_MAJOR >= 3
@@ -81,23 +50,27 @@ public:
   int ReleaseData() override;
 
 protected:
-  AmrParticleDataAdaptor() = default;
-  ~AmrParticleDataAdaptor() = default;
+  AmrParticleDataAdaptor()
+  {
+    m_meshAdaptor = AmrDataAdaptor::New();
+    m_particleAdaptor = ParticleDataAdaptor<NStructReal, NStructInt, NArrayReal, NArrayInt>::New();
+  }
+
+  ~AmrParticleDataAdaptor()
+  {
+    m_meshAdaptor->Delete();
+    m_particleAdaptor->Delete();
+  }
 
 private:
-  vtkPolyData* BuildParticles();
+  ParticleDataAdaptor<NStructReal, NStructInt, NArrayReal, NArrayInt>* m_particleAdaptor;
+  AmrDataAdaptor* m_meshAdaptor;
 
+  const std::string m_meshName = "mesh";
   const std::string m_particlesName = "particles";
-  amrex::ParticleContainer<NStructReal, NStructInt, NArrayReal, NArrayInt>* m_particles;
-  std::map<std::string, std::vector<int>> m_realStructs;
-  std::map<std::string, int> m_intStructs;
-  std::map<std::string, std::vector<int>> m_realArrays;
-  std::map<std::string, int> m_intArrays;
 };
-
 }
 
 #include "AMReX_AmrParticleDataAdaptorI.H"
-
 #endif // AMREX_PARTICLES
 #endif // AMReX_AmrParticleDataAdaptor_h
